Lets symlink check if file exists.

+++ symlink.rb
@@ -8,5 +8,6 @@ Dir['*'].each do |file|
   next if file =~ /install/
   target = File.join(home, ".#{file}")
   next if File.symlink? target
+  next if File.exists? target
   `ln -s #{File.expand_path file} #{target}`
 end
